--- components/_util/formItemStatus.ts
@@ -1,9 +1,0 @@
-import React from 'react';
-import { ValidateStatus } from '../form/FormItem';
-
-export interface FormItemStatusContextProps {
-  status?: ValidateStatus;
-  hasFeedback?: boolean;
-}
-
-export const FormItemStatusContext = React.createContext<FormItemStatusContextProps>({});
