+++ components/_util/getFeedbackIcon.tsx
@@ -1,6 +1,16 @@
 import React from 'react';
+import CheckCircleFilled from '@ant-design/icons/CheckCircleFilled';
+import ExclamationCircleFilled from '@ant-design/icons/ExclamationCircleFilled';
+import CloseCircleFilled from '@ant-design/icons/CloseCircleFilled';
+import LoadingOutlined from '@ant-design/icons/LoadingOutlined';
 import { ValidateStatus } from '../form/FormItem';
-import iconMap from './validationIcons';
+
+const iconMap = {
+  success: CheckCircleFilled,
+  warning: ExclamationCircleFilled,
+  error: CloseCircleFilled,
+  validating: LoadingOutlined,
+};
 
 const getFeedbackIcon = (prefixCls: string, status?: ValidateStatus) => {
   const IconNode = status && iconMap[status];
