+++ components/input/TextArea.tsx
@@ -9,9 +9,9 @@ import { ConfigContext } from '../config-provider';
 import { fixControlledValue, resolveOnChange, triggerFocus, InputFocusOptions } from './Input';
 import SizeContext, { SizeType } from '../config-provider/SizeContext';
 import { ValidateStatus } from '../form/FormItem';
-import { FormItemStatusContext } from '../_util/formItemStatus';
 import getStatusClassNames from '../_util/getStatusClassNames';
 import getFeedbackIcon from '../_util/getFeedbackIcon';
+import { FormItemStatusContext } from '../form/context';
 
 interface ShowCountProps {
   formatter: (args: { count: number; maxLength?: number }) => string;
